dom/media/tests/mochitest/test_getUserMedia_constraints.html

<!DOCTYPE HTML>
<html>
<head>
  <script src="mediaStreamPlayback.js"></script>
  <script src="constraints.js"></script>
</head>
<body>
<pre id="test">
<script type="application/javascript">
createHTML({ title: "Test getUserMedia constraints", bug: "882145" });
/**
  Tests covering gUM constraints API for audio, video and fake video. Exercise
  successful parsing code and ensure that unknown required constraints and
  overconstraining cases produce appropriate errors.
*/
var tests = [
  // Each test here tests a different constraint or codepath.
  { message: "unknown required constraint on video ignored",
    constraints: { video: { somethingUnknown: { exact: 0 } },
                   fake: true },
    error: null },
  { message: "unknown required constraint on audio ignored",
    constraints: { audio: { somethingUnknown: { exact: 0 } },
                   fake: true },
    error: null },
  { message: "video overconstrained by facingMode fails",
    constraints: { video: { facingMode:{ exact: 'left' } },
                   fake: true },
    error: "NotFoundError" },
  { message: "video overconstrained by facingMode array fails",
    constraints: { video: { facingMode:{ exact: ['left', 'right'] } },
                   fake: true },
    error: "NotFoundError" },
  { message: "audio overconstrained by facingMode ignored",
    constraints: { audio: { facingMode: { exact: 'left' } },
                   fake: true },
    error: null },
  { message: "full screensharing requires permission",
    constraints: { video: { mediaSource: 'screen' } },
    error: "PermissionDeniedError" },
  { message: "application screensharing requires permission",
    constraints: { video: { mediaSource: 'application' } },
    error: "PermissionDeniedError" },
  { message: "window screensharing requires permission",
    constraints: { video: { mediaSource: 'window' } },
    error: "PermissionDeniedError" },
  { message: "browser screensharing requires permission",
    constraints: { video: { mediaSource: 'browser' } },
    error: "PermissionDeniedError" },
  { message: "unknown mediaSource fails",
    constraints: { video: { mediaSource: 'uncle' } },
    error: "NotFoundError" },
  { message: "Success-path: optional video facingMode + audio ignoring facingMode",
    constraints: { fake: true,
                   audio: { mediaSource: 'microphone',
                            facingMode: 'left',
                            foo: 0,
                            advanced: [{ facingMode: 'environment' },
                                       { facingMode: 'user' },
                                       { bar: 0 }] },
                   video: { mediaSource: 'camera',
                            foo: 0,
                            advanced: [{ facingMode: 'environment' },
                                       { facingMode: ['user'] },
                                       { facingMode: ['left', 'right', 'user'] },
                                       { bar: 0 }] } },
    error: null },
  { message: "legacy facingMode ignored",
    constraints: { video: { mandatory: { facingMode: 'left' } }, fake: true },
    error: null },
];

/**
 * Starts the test run by running through each constraint
 * test by verifying that the right resolution and rejection is fired.
 */

runTest(function() {
  var p = new Promise(resolve => SpecialPowers.pushPrefEnv({
      set : [ ['media.getusermedia.browser.enabled', false],
              ['media.getusermedia.screensharing.enabled', false] ]
    }, resolve));

  tests.reduce((p, test) =>
    p.then(() => navigator.mediaDevices.getUserMedia(test.constraints))
    .then(() => is(null, test.error, test.message),
          e => is(e.name, test.error, test.message + ": " + e.message)), p)
  .catch(e => ok(false, "Unexpected failure: " + e.name + ", " + e.message))
  .then(SimpleTest.finish);
});

</script>
</pre>
</body>
</html>
